Chore: added comments and moved TextBlockUtils to its own file
